+++ dombusgateway/CHANGELOG.md
@@ -6,6 +6,9 @@ Developed by Creasol - https://www.creasol.it/domotics
 
 Below the changes related to the addon only
 
+## [0.4.1] 2025-11-30
+* Several new features: please check the CHANGELOG in the DomBusGateway GitHub repository.
+
 ## [0.3.3] 2025-10-09
 * Added --data_dir parameter to override dataDir (persistent data folder) in dombusgateway_conf.py
 * Devices and Modules dictionaries are saved after 15s since changes, to avoid loosing data when container is restarted